requirements.txt
python-pptx==0.6.23
pandas==2.1.4
matplotlib==3.8.2
openpyxl==3.1.2
schedule==1.2.1
Pillow==10.1.0
reportlab==4.0.8
PyYAML==6.0.1

# Optional: For additional data sources
requests>=2.31.0        # REST API support
sqlalchemy>=2.0.0       # SQL database support
gspread>=5.12.0         # Google Sheets support
google-auth>=2.25.0     # Google authentication
